+++ tests/Models/AlternativeStatusModel.php
@@ -4,6 +4,6 @@ namespace Spatie\ModelStatus\Tests\Models;
 
 use Spatie\ModelStatus\Status;
 
-class StatusTestModel extends Status
+class AlternativeStatusModel extends Status
 {
 }