+++ tests/Models/InvalidStatusModel.php
@@ -4,6 +4,6 @@ namespace Spatie\ModelStatus\Tests\Models;
 
 use Illuminate\Database\Eloquent\Model;
 
-class InvalidExtendTestModel extends Model
+class InvalidStatusModel extends Model
 {
 }