+++ tests/TestCase.php
@@ -2,8 +2,10 @@
 
 namespace Spatie\ModelStatus\Tests;
 
+use CreateStatusesTable;
 use Illuminate\Database\Schema\Blueprint;
 use Orchestra\Testbench\TestCase as BaseTestCase;
+use Spatie\ModelStatus\ModelStatusServiceProvider;
 
 class TestCase extends BaseTestCase
 {
@@ -14,14 +16,11 @@ class TestCase extends BaseTestCase
         $this->setUpDatabase();
     }
 
-    public function getEnvironmentSetUp($app)
+    protected function getPackageProviders($app)
     {
-        $app['config']->set(
-            'model-status.status_model',
-            \Spatie\ModelStatus\Tests\Models\StatusTestModel::class
-        );
-
-        $app['config']->set('app.key', '6rE9Nz59bGRbeMATftriyQjrpF7DcOQm');
+        return [
+            ModelStatusServiceProvider::class,
+        ];
     }
 
     protected function setUpDatabase()
@@ -40,6 +39,6 @@ class TestCase extends BaseTestCase
 
         include_once __DIR__.'/../database/migrations/create_statuses_table.php.stub';
 
-        (new \CreateStatusesTable())->up();
+        (new CreateStatusesTable())->up();
     }
 }