+++ CHANGELOG.md
@@ -2,10 +2,14 @@
 
 All notable changes to `laravel-model-status` will be documented in this file
 
+## 1.0.6 - 2018-02-09
+
+- add `__toString` to `Spatie\ModelStatus\Status`
+- add `currentStatus` scope
+
 ## 1.0.5 - 2018-02-09
 
-- add return string when called as an attribute
-- add `hasStatus('status')` to find the models with the given status as there current status
+**THIS VERSION CONTAINS BUGS, DO NOT USE**
 
 ## 1.0.4 - 2018-02-08